Extract `TDRClient` from azul.terra module
